Move extension stuff to registry
