+++ src/registry/extensions/projectExplorer/revealInFileExplorer.ts
